fix: replace [source] only when stand-alone on line
Many field references being added to docs as a part of the ECS effort
contain the literal string `[source]`, so we need to be more conservative
about which instances of `[source]` we replace with `[source,shell]`.

By matching a regular expression instead of a substring, we can match
only the instances of `[source]` that are stand-alone on a line.